+++ services/review/internal/service/service.go
@@ -8,15 +8,15 @@ import (
 )
 
 type ReviewRepository interface {
-	GetReviews(ctx context.Context, restaurantID int32, page, limit int) ([]models.Review, int, error)
+	GetReviews(ctx context.Context, restaurantID int64, page, limit int) ([]models.Review, int, error)
 	CreateReview(ctx context.Context, review models.Review) (models.Review, error)
 	UpdateReview(ctx context.Context, review models.UpdateReview) (models.Review, error)
-	DeleteReview(ctx context.Context, reviewID int32, userID int32) error
-	GetRestaurantStats(ctx context.Context, restaurantID int32) (models.RestaurantStats, error)
+	DeleteReview(ctx context.Context, reviewID int32, userID int64) error
+	GetRestaurantStats(ctx context.Context, restaurantID int64) (models.RestaurantStats, error)
 }
 
 type RestaurantClient interface {
-	GetRestaurant(ctx context.Context, restaurantID int32) (models.Restaurant, error)
+	GetRestaurant(ctx context.Context, restaurantID int64) (models.Restaurant, error)
 }
 
 type Service struct {
@@ -28,7 +28,7 @@ func New(repo ReviewRepository, restaurantClient RestaurantClient) *Service {
 	return &Service{repo: repo, restaurantClient: restaurantClient}
 }
 
-func (s *Service) GetReviews(ctx context.Context, restaurantID int32, page, limit int) (models.ReviewsPage, error) {
+func (s *Service) GetReviews(ctx context.Context, restaurantID int64, page, limit int) (models.ReviewsPage, error) {
 	if restaurantID <= 0 {
 		slog.Error("Restaurant ID is not valid: value must be a positive integer", "restaurantID", restaurantID)
 		return models.ReviewsPage{}, models.ErrInvalidRestaurantID
@@ -38,7 +38,7 @@ func (s *Service) GetReviews(ctx context.Context, restaurantID int32, page, limi
 		return models.ReviewsPage{}, models.ErrLimitTooLarge
 	}
 
-	restaurant, err := s.restaurantClient.GetRestaurant(ctx, int32(restaurantID))
+	restaurant, err := s.restaurantClient.GetRestaurant(ctx, restaurantID)
 	if err != nil {
 		slog.Error("Failed to get restaurant details", "restaurantID", restaurantID, "error", err)
 		return models.ReviewsPage{}, err
@@ -114,7 +114,7 @@ func (s *Service) DeleteReview(ctx context.Context, deleteReq models.DeleteRevie
 	return s.repo.DeleteReview(ctx, deleteReq.ReviewID, deleteReq.UserID)
 }
 
-func (s *Service) GetRestaurantStats(ctx context.Context, restaurantID int32) (models.RestaurantStats, error) {
+func (s *Service) GetRestaurantStats(ctx context.Context, restaurantID int64) (models.RestaurantStats, error) {
 	if restaurantID <= 0 {
 		slog.Error("Restaurant ID is not valid: value must be a positive integer", "restaurantID", restaurantID)
 		return models.RestaurantStats{}, models.ErrInvalidRestaurantID
